Added Support for onChange when not in view
Earlier it was working fine if isVisible was true and doing the desired actions but if the component goes out of the view area (isVisible == false) it wasn’t triggering the action ( not returning the false value in onChange). So I added another condition that it will still send the isVisible boolean in props onChange even if it is not in the view area. Useful when you want to pause a video when not in user view.